[Quartermaster] refactor: Split SpellsPanel into partial classes (#778)
SpellsPanel.axaml.cs split into 4 focused files:
- SpellsPanel.axaml.cs (394 lines) - Core/initialization
- SpellsPanel.DataLoading.cs (302 lines) - Data loading
- SpellsPanel.EventHandlers.cs (182 lines) - Event handlers
- SpellsPanel.SummaryBuilders.cs (453 lines) - Summary UI builders

Main file reduced from 1299 to 394 lines (70% reduction).
