+++ Quartermaster/Quartermaster/Views/Panels/SpellsPanel.axaml.cs
@@ -2,7 +2,6 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 using Avalonia.Media;
-using Avalonia.Media.Imaging;
 using Avalonia.Styling;
 using Quartermaster.Services;
 using Quartermaster.ViewModels;
@@ -152,25 +151,6 @@ public partial class SpellsPanel : UserControl
         }
     }
 
-    private void OnClassRadioChecked(object? sender, Avalonia.Interactivity.RoutedEventArgs e)
-    {
-        if (_isLoading) return;
-
-        if (sender is RadioButton radio && radio.IsChecked == true)
-        {
-            var index = Array.IndexOf(_classRadios, radio);
-            if (index >= 0 && index != _selectedClassIndex)
-            {
-                _isLoading = true;
-                _selectedClassIndex = index;
-                // Reload spells for the selected class
-                if (_currentCreature != null)
-                    LoadSpellsForClass(_selectedClassIndex);
-                _isLoading = false;
-            }
-        }
-    }
-
     /// <summary>
     /// Sets the display service for 2DA/TLK lookups.
     /// </summary>
@@ -186,448 +166,6 @@ public partial class SpellsPanel : UserControl
     {
         _itemIconService = iconService;
         UnifiedLogger.LogUI(LogLevel.INFO, $"SpellsPanel: IconService set, IsGameDataAvailable={iconService?.IsGameDataAvailable}");
-    }
-
-    public void LoadCreature(UtcFile? creature)
-    {
-        _displayedSpells.Clear();
-        _allSpells.Clear();
-        _knownSpellIds.Clear();
-        _memorizedSpellIds.Clear();
-        _currentCreature = creature;
-        _selectedClassIndex = 0;
-
-        if (creature == null)
-        {
-            ClearPanel();
-            return;
-        }
-
-        // Prevent dirty marking during load
-        _isLoading = true;
-
-        // Show loading state
-        if (_loadingText != null)
-            _loadingText.IsVisible = true;
-
-        // Update class radio buttons based on creature's classes
-        UpdateClassRadioButtons(creature);
-
-        // Load spells for the first (selected) class
-        LoadSpellsForClass(_selectedClassIndex);
-
-        // Hide loading state
-        if (_loadingText != null)
-            _loadingText.IsVisible = false;
-
-        _isLoading = false;
-    }
-
-    private void UpdateClassRadioButtons(UtcFile creature)
-    {
-        // Reset all radio buttons
-        bool foundCaster = false;
-
-        for (int i = 0; i < _classRadios.Length; i++)
-        {
-            var radio = _classRadios[i];
-            if (radio == null) continue;
-
-            if (i < creature.ClassList.Count)
-            {
-                var classEntry = creature.ClassList[i];
-                var className = _displayService?.GetClassName(classEntry.Class) ?? $"Class {classEntry.Class}";
-
-                // Check if this class can cast spells using the display service
-                bool isCaster = _displayService?.IsCasterClass(classEntry.Class) ?? false;
-                int maxSpellLevel = isCaster ? (_displayService?.GetMaxSpellLevel(classEntry.Class, classEntry.ClassLevel) ?? -1) : -1;
-
-                // Check if this class actually has spells in the creature data
-                bool hasSpellsInData = classEntry.KnownSpells.Any(list => list.Count > 0) ||
-                                       classEntry.MemorizedSpells.Any(list => list.Count > 0);
-
-                // Format: "Wizard (10) - Lvl 5" or "Fighter (5)" for non-casters
-                if (isCaster && maxSpellLevel >= 0)
-                {
-                    radio.Content = $"{className} ({classEntry.ClassLevel}) - Lvl {maxSpellLevel}";
-                }
-                else if (isCaster || hasSpellsInData)
-                {
-                    // Caster class or has spell data
-                    radio.Content = $"{className} ({classEntry.ClassLevel})";
-                }
-                else
-                {
-                    radio.Content = $"{className} ({classEntry.ClassLevel})";
-                }
-
-                // Enable if: detected as caster with spells, OR has actual spell data in creature
-                radio.IsEnabled = (isCaster && maxSpellLevel >= 0) || hasSpellsInData;
-                radio.IsVisible = true;
-
-                // Select first enabled caster class by default
-                if (!foundCaster && radio.IsEnabled)
-                {
-                    radio.IsChecked = true;
-                    _selectedClassIndex = i;
-                    foundCaster = true;
-                }
-            }
-            else
-            {
-                radio.Content = $"Class {i + 1}";
-                radio.IsEnabled = false;
-                radio.IsVisible = i < 3; // Show first 3 placeholders
-            }
-        }
-
-        // If no caster class found, select first anyway (even if disabled)
-        if (!foundCaster && _classRadios[0] != null)
-            _classRadios[0].IsChecked = true;
-    }
-
-    private void LoadSpellsForClass(int classIndex)
-    {
-        _displayedSpells.Clear();
-        _allSpells.Clear();
-        _knownSpellIds.Clear();
-        _memorizedSpellIds.Clear();
-        _isSpontaneousCaster = false;
-
-        if (_currentCreature == null || _displayService == null)
-        {
-            ApplyFilters();
-            UpdateSummary();
-            return;
-        }
-
-        if (classIndex >= _currentCreature.ClassList.Count)
-        {
-            ApplyFilters();
-            UpdateSummary();
-            return;
-        }
-
-        var classEntry = _currentCreature.ClassList[classIndex];
-
-        // Check if this is a spontaneous caster (Sorcerer, Bard)
-        _isSpontaneousCaster = _displayService.IsSpontaneousCaster(classEntry.Class);
-
-        // Populate known spell IDs from parsed KnownList0-9
-        for (int level = 0; level < 10; level++)
-        {
-            foreach (var spell in classEntry.KnownSpells[level])
-            {
-                _knownSpellIds.Add(spell.Spell);
-            }
-        }
-
-        // Populate memorized spell IDs from parsed MemorizedList0-9
-        for (int level = 0; level < 10; level++)
-        {
-            foreach (var spell in classEntry.MemorizedSpells[level])
-            {
-                _memorizedSpellIds.Add(spell.Spell);
-            }
-        }
-
-        // Load all spells from spells.2da
-        LoadAllSpells(classEntry.Class);
-
-        // Apply filters
-        ApplyFilters();
-
-        // Update summary
-        UpdateSummary();
-    }
-
-    private void LoadAllSpells(int classId)
-    {
-        if (_displayService == null) return;
-
-        var allSpellIds = _displayService.GetAllSpellIds();
-
-        foreach (var spellId in allSpellIds)
-        {
-            var vm = CreateSpellViewModel(spellId, classId);
-            if (vm != null)
-                _allSpells.Add(vm);
-        }
-
-        // Sort by name
-        _allSpells = _allSpells.OrderBy(s => s.SpellName).ToList();
-    }
-
-    private SpellListViewModel? CreateSpellViewModel(int spellId, int classId)
-    {
-        if (_displayService == null) return null;
-
-        var spellName = _displayService.GetSpellName(spellId);
-        var spellInfo = _displayService.GetSpellInfo(spellId);
-
-        if (spellInfo == null)
-        {
-            // Basic fallback if no spell info available
-            var fallbackVm = new SpellListViewModel
-            {
-                SpellId = spellId,
-                SpellName = spellName,
-                SpellLevel = 0,
-                SpellLevelDisplay = "?",
-                InnateLevel = 0,
-                InnateLevelDisplay = "?",
-                School = SpellSchool.Unknown,
-                SchoolName = "Unknown",
-                IsKnown = _knownSpellIds.Contains(spellId),
-                IsMemorized = _memorizedSpellIds.Contains(spellId),
-                IsBlocked = false,
-                IsSpontaneousCaster = _isSpontaneousCaster,
-                BlockedReason = "",
-                Description = spellName
-            };
-            fallbackVm.OnKnownChanged = OnSpellKnownChanged;
-            fallbackVm.OnMemorizedChanged = OnSpellMemorizedChanged;
-            LoadSpellIcon(fallbackVm);
-            return fallbackVm;
-        }
-
-        // Get spell level for this class
-        int spellLevel = spellInfo.GetLevelForClass(classId);
-        bool isAvailableToClass = spellLevel >= 0;
-
-        var isKnown = _knownSpellIds.Contains(spellId);
-        var isMemorized = _memorizedSpellIds.Contains(spellId);
-        var isBlocked = !isAvailableToClass;
-        var blockedReason = isBlocked ? "Not available to this class" : "";
-
-        // Determine status display
-        string statusText;
-        IBrush statusColor;
-        IBrush rowBackground;
-        double textOpacity;
-
-        if (isBlocked)
-        {
-            statusText = "Blocked";
-            statusColor = GetDisabledBrush();
-            rowBackground = GetTransparentRowBackground(statusColor, 20);
-            textOpacity = 0.5;
-        }
-        else if (isMemorized)
-        {
-            statusText = "Memorized";
-            statusColor = GetSelectionBrush();
-            rowBackground = GetTransparentRowBackground(statusColor, 30);
-            textOpacity = 1.0;
-        }
-        else if (isKnown)
-        {
-            statusText = "Known";
-            statusColor = GetSuccessBrush();
-            rowBackground = GetTransparentRowBackground(statusColor, 30);
-            textOpacity = 1.0;
-        }
-        else
-        {
-            statusText = "";
-            statusColor = Brushes.Transparent;
-            rowBackground = Brushes.Transparent;
-            textOpacity = 0.7;
-        }
-
-        var vm = new SpellListViewModel
-        {
-            SpellId = spellId,
-            SpellName = spellName,
-            SpellLevel = spellLevel >= 0 ? spellLevel : spellInfo.InnateLevel,
-            SpellLevelDisplay = spellLevel >= 0 ? spellLevel.ToString() : "-",
-            InnateLevel = spellInfo.InnateLevel,
-            InnateLevelDisplay = spellInfo.InnateLevel.ToString(),
-            School = spellInfo.School,
-            SchoolName = GetSchoolName(spellInfo.School),
-            IsKnown = isKnown,
-            IsMemorized = isMemorized,
-            IsBlocked = isBlocked,
-            IsSpontaneousCaster = _isSpontaneousCaster,
-            BlockedReason = blockedReason,
-            Description = BuildTooltip(spellName, spellInfo, blockedReason),
-            StatusText = statusText,
-            StatusColor = statusColor,
-            RowBackground = rowBackground,
-            TextOpacity = textOpacity
-        };
-
-        // Wire up change handlers
-        vm.OnKnownChanged = OnSpellKnownChanged;
-        vm.OnMemorizedChanged = OnSpellMemorizedChanged;
-
-        // Load spell icon if available
-        LoadSpellIcon(vm);
-
-        return vm;
-    }
-
-    /// <summary>
-    /// Loads the game icon for a spell from spells.2da IconResRef.
-    /// Icons are loaded lazily when binding requests them.
-    /// </summary>
-    private void LoadSpellIcon(SpellListViewModel spellVm)
-    {
-        // Don't load upfront - use lazy loading via IconBitmap getter
-        // This prevents loading 467+ bitmaps at once which crashes Avalonia
-        spellVm.SetIconService(_itemIconService);
-    }
-
-    private void OnSpellKnownChanged(SpellListViewModel spell, bool isNowKnown)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        if (_selectedClassIndex >= _currentCreature.ClassList.Count) return;
-
-        var classEntry = _currentCreature.ClassList[_selectedClassIndex];
-
-        if (isNowKnown)
-        {
-            // Add to known spells
-            if (!_knownSpellIds.Contains(spell.SpellId))
-            {
-                _knownSpellIds.Add(spell.SpellId);
-
-                // Add to model at appropriate spell level
-                classEntry.KnownSpells[spell.SpellLevel].Add(new KnownSpell
-                {
-                    Spell = (ushort)spell.SpellId,
-                    SpellFlags = 0x01,
-                    SpellMetaMagic = 0
-                });
-            }
-        }
-        else
-        {
-            // Remove from known spells
-            _knownSpellIds.Remove(spell.SpellId);
-
-            // Remove from model
-            var knownList = classEntry.KnownSpells[spell.SpellLevel];
-            var existing = knownList.FirstOrDefault(s => s.Spell == spell.SpellId);
-            if (existing != null)
-            {
-                knownList.Remove(existing);
-            }
-
-            // Also remove from memorized if it was memorized
-            if (_memorizedSpellIds.Contains(spell.SpellId))
-            {
-                _memorizedSpellIds.Remove(spell.SpellId);
-                var memorizedList = classEntry.MemorizedSpells[spell.SpellLevel];
-                var memorized = memorizedList.FirstOrDefault(s => s.Spell == spell.SpellId);
-                if (memorized != null)
-                {
-                    memorizedList.Remove(memorized);
-                }
-                spell.IsMemorized = false;
-            }
-        }
-
-        // Update visual status
-        UpdateSpellVisualStatus(spell);
-
-        // Notify that spells changed
-        SpellsChanged?.Invoke(this, EventArgs.Empty);
-
-        // Update summary
-        UpdateSummary();
-    }
-
-    private void OnSpellMemorizedChanged(SpellListViewModel spell, bool isNowMemorized)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        if (_selectedClassIndex >= _currentCreature.ClassList.Count) return;
-
-        var classEntry = _currentCreature.ClassList[_selectedClassIndex];
-
-        if (isNowMemorized)
-        {
-            // Add to memorized spells
-            if (!_memorizedSpellIds.Contains(spell.SpellId))
-            {
-                _memorizedSpellIds.Add(spell.SpellId);
-
-                // Add to model at appropriate spell level
-                classEntry.MemorizedSpells[spell.SpellLevel].Add(new MemorizedSpell
-                {
-                    Spell = (ushort)spell.SpellId,
-                    SpellFlags = 0x01,
-                    SpellMetaMagic = 0,
-                    Ready = 1
-                });
-            }
-        }
-        else
-        {
-            // Remove from memorized spells
-            _memorizedSpellIds.Remove(spell.SpellId);
-
-            // Remove from model
-            var memorizedList = classEntry.MemorizedSpells[spell.SpellLevel];
-            var existing = memorizedList.FirstOrDefault(s => s.Spell == spell.SpellId);
-            if (existing != null)
-            {
-                memorizedList.Remove(existing);
-            }
-        }
-
-        // Update visual status
-        UpdateSpellVisualStatus(spell);
-
-        // Notify that spells changed
-        SpellsChanged?.Invoke(this, EventArgs.Empty);
-
-        // Update summary
-        UpdateSummary();
-    }
-
-    private void UpdateSpellVisualStatus(SpellListViewModel spell)
-    {
-        // Update status based on new known/memorized state
-        bool isKnown = spell.IsKnown;
-        bool isMemorized = _memorizedSpellIds.Contains(spell.SpellId);
-
-        if (spell.IsBlocked)
-        {
-            spell.StatusText = "Blocked";
-            spell.StatusColor = GetDisabledBrush();
-            spell.RowBackground = GetTransparentRowBackground(spell.StatusColor, 20);
-            spell.TextOpacity = 0.5;
-        }
-        else if (isKnown && isMemorized)
-        {
-            spell.StatusText = "K + M";
-            spell.StatusColor = GetSelectionBrush();
-            spell.RowBackground = GetTransparentRowBackground(spell.StatusColor, 30);
-            spell.TextOpacity = 1.0;
-        }
-        else if (isMemorized)
-        {
-            // Memorized but not known (edge case - shouldn't happen normally)
-            spell.StatusText = "Memorized";
-            spell.StatusColor = GetSelectionBrush();
-            spell.RowBackground = GetTransparentRowBackground(spell.StatusColor, 30);
-            spell.TextOpacity = 1.0;
-        }
-        else if (isKnown)
-        {
-            spell.StatusText = "Known";
-            spell.StatusColor = GetSuccessBrush();
-            spell.RowBackground = GetTransparentRowBackground(spell.StatusColor, 30);
-            spell.TextOpacity = 1.0;
-        }
-        else
-        {
-            spell.StatusText = "";
-            spell.StatusColor = Brushes.Transparent;
-            spell.RowBackground = Brushes.Transparent;
-            spell.TextOpacity = 0.7;
-        }
     }
 
     private static string GetSchoolName(SpellSchool school)
@@ -735,449 +273,6 @@ public partial class SpellsPanel : UserControl
             _noSpellsText.IsVisible = _displayedSpells.Count == 0;
 
         UpdateSummary();
-    }
-
-    private void UpdateSummary()
-    {
-        // Update all caster classes summary
-        UpdateAllClassesSummary();
-
-        // Update known spells list
-        UpdateKnownSpellsList();
-
-        // Update selected class spell slot summary
-        UpdateSpellSlotSummary();
-    }
-
-    /// <summary>
-    /// Updates the spell slot table on the left side.
-    /// Shows a grid with spell levels as rows and caster classes as columns.
-    /// </summary>
-    private void UpdateAllClassesSummary()
-    {
-        if (_spellSlotTableGrid == null || _spellSlotTableBorder == null) return;
-
-        // Clear existing content
-        _spellSlotTableGrid.Children.Clear();
-        _spellSlotTableGrid.ColumnDefinitions.Clear();
-        _spellSlotTableGrid.RowDefinitions.Clear();
-
-        if (_currentCreature == null || _displayService == null)
-        {
-            _spellSlotTableBorder.IsVisible = false;
-            return;
-        }
-
-        // Get theme-aware font sizes
-        var normalFontSize = this.FindResource("FontSizeNormal") as double? ?? 14;
-        var smallFontSize = this.FindResource("FontSizeSmall") as double? ?? 12;
-
-        var casterClasses = new List<(int classIndex, string className, int classLevel, int[] limits, CreatureClass classEntry, bool isSpontaneous)>();
-
-        // Find all caster classes
-        for (int i = 0; i < _currentCreature.ClassList.Count; i++)
-        {
-            var classEntry = _currentCreature.ClassList[i];
-            var className = _displayService.GetClassName(classEntry.Class) ?? $"Class {classEntry.Class}";
-            var isSpontaneous = _displayService.IsSpontaneousCaster(classEntry.Class);
-
-            // For spontaneous casters, use spells known limit; for prepared casters, use spell slots
-            var limits = isSpontaneous
-                ? _displayService.GetSpellsKnownLimit(classEntry.Class, classEntry.ClassLevel)
-                : _displayService.GetSpellSlots(classEntry.Class, classEntry.ClassLevel);
-
-            // Check if this class has spell slots/limits or has spell data
-            bool hasSpellLimits = limits != null && limits.Any(s => s > 0);
-            bool hasSpellData = classEntry.KnownSpells.Any(list => list.Count > 0) ||
-                                classEntry.MemorizedSpells.Any(list => list.Count > 0);
-
-            if (hasSpellLimits || hasSpellData)
-            {
-                casterClasses.Add((i, className, classEntry.ClassLevel, limits ?? new int[10], classEntry, isSpontaneous));
-            }
-        }
-
-        if (casterClasses.Count == 0)
-        {
-            _spellSlotTableBorder.IsVisible = false;
-            return;
-        }
-
-        _spellSlotTableBorder.IsVisible = true;
-
-        // Find all spell levels that have limits across any class
-        var activeLevels = new List<int>();
-        for (int level = 0; level <= 9; level++)
-        {
-            if (casterClasses.Any(c => c.limits[level] > 0))
-            {
-                activeLevels.Add(level);
-            }
-        }
-
-        if (activeLevels.Count == 0)
-        {
-            _spellSlotTableBorder.IsVisible = false;
-            return;
-        }
-
-        // Build grid structure: first column for level labels, then one column per class
-        // Row 0 = header row with class names
-        // Rows 1+ = spell levels
-
-        // Add column definitions: Label column + one per class
-        _spellSlotTableGrid.ColumnDefinitions.Add(new ColumnDefinition(GridLength.Auto));
-        foreach (var _ in casterClasses)
-        {
-            _spellSlotTableGrid.ColumnDefinitions.Add(new ColumnDefinition(new GridLength(1, GridUnitType.Star)));
-        }
-
-        // Add row definitions: header + one per active level
-        _spellSlotTableGrid.RowDefinitions.Add(new RowDefinition(GridLength.Auto)); // Header
-        foreach (var _ in activeLevels)
-        {
-            _spellSlotTableGrid.RowDefinitions.Add(new RowDefinition(GridLength.Auto));
-        }
-
-        // Add header row - "Lvl" label in first column
-        var lvlHeader = new TextBlock
-        {
-            Text = "Lvl",
-            FontWeight = Avalonia.Media.FontWeight.SemiBold,
-            FontSize = smallFontSize,
-            Margin = new Avalonia.Thickness(0, 0, 10, 4),
-            HorizontalAlignment = Avalonia.Layout.HorizontalAlignment.Center
-        };
-        Grid.SetRow(lvlHeader, 0);
-        Grid.SetColumn(lvlHeader, 0);
-        _spellSlotTableGrid.Children.Add(lvlHeader);
-
-        // Add class name headers
-        for (int col = 0; col < casterClasses.Count; col++)
-        {
-            var (classIndex, className, classLevel, _, _, _) = casterClasses[col];
-            var isSelected = classIndex == _selectedClassIndex;
-
-            var classHeader = new TextBlock
-            {
-                Text = className,
-                FontWeight = Avalonia.Media.FontWeight.SemiBold,
-                FontSize = smallFontSize,
-                Margin = new Avalonia.Thickness(4, 0, 4, 4),
-                HorizontalAlignment = Avalonia.Layout.HorizontalAlignment.Center,
-                Foreground = isSelected
-                    ? GetInfoBrush()
-                    : this.FindResource("SystemControlForegroundBaseHighBrush") as IBrush
-                      ?? GetDisabledBrush()
-            };
-            Grid.SetRow(classHeader, 0);
-            Grid.SetColumn(classHeader, col + 1);
-            _spellSlotTableGrid.Children.Add(classHeader);
-        }
-
-        // Add data rows for each spell level
-        for (int rowIdx = 0; rowIdx < activeLevels.Count; rowIdx++)
-        {
-            int spellLevel = activeLevels[rowIdx];
-            int gridRow = rowIdx + 1; // +1 for header
-
-            // Level label
-            var levelLabel = new TextBlock
-            {
-                Text = spellLevel.ToString(),
-                FontSize = smallFontSize,
-                FontWeight = Avalonia.Media.FontWeight.SemiBold,
-                Margin = new Avalonia.Thickness(0, 2, 10, 2),
-                HorizontalAlignment = Avalonia.Layout.HorizontalAlignment.Center,
-                VerticalAlignment = Avalonia.Layout.VerticalAlignment.Center
-            };
-            Grid.SetRow(levelLabel, gridRow);
-            Grid.SetColumn(levelLabel, 0);
-            _spellSlotTableGrid.Children.Add(levelLabel);
-
-            // Slot/known counts for each class
-            for (int col = 0; col < casterClasses.Count; col++)
-            {
-                var (classIndex, className, _, limits, classEntry, isSpontaneous) = casterClasses[col];
-                int totalLimit = limits[spellLevel];
-
-                // Only count spells that are actual class spells (not feat-based abilities like Barbarian Rage)
-                // Filter to spells that have a valid level for this class in spells.2da
-                int usedCount = 0;
-                foreach (var knownSpell in classEntry.KnownSpells[spellLevel])
-                {
-                    var spellInfo = _displayService.GetSpellInfo(knownSpell.Spell);
-                    int classSpellLevel = spellInfo?.GetLevelForClass(classEntry.Class) ?? -1;
-
-                    // Only count spells that are valid for this class (filter out feat-based abilities)
-                    if (spellInfo != null && classSpellLevel >= 0)
-                    {
-                        usedCount++;
-                    }
-                }
-
-                var isSelected = classIndex == _selectedClassIndex;
-
-                string cellText;
-                IBrush cellColor;
-
-                if (totalLimit <= 0)
-                {
-                    cellText = "-";
-                    cellColor = GetDisabledBrush();
-                }
-                else if (usedCount >= totalLimit)
-                {
-                    // Full - show in gold/yellow
-                    cellText = $"{usedCount}/{totalLimit}";
-                    cellColor = GetSelectionBrush();
-                }
-                else if (usedCount > 0)
-                {
-                    // Partial - show in green
-                    cellText = $"{usedCount}/{totalLimit}";
-                    cellColor = GetSuccessBrush();
-                }
-                else
-                {
-                    // Empty - normal color
-                    cellText = $"0/{totalLimit}";
-                    cellColor = isSelected
-                        ? GetInfoBrush()
-                        : this.FindResource("SystemControlForegroundBaseMediumBrush") as IBrush
-                          ?? GetDisabledBrush();
-                }
-
-                var slotCell = new TextBlock
-                {
-                    Text = cellText,
-                    FontSize = smallFontSize,
-                    Margin = new Avalonia.Thickness(4, 2, 4, 2),
-                    HorizontalAlignment = Avalonia.Layout.HorizontalAlignment.Center,
-                    VerticalAlignment = Avalonia.Layout.VerticalAlignment.Center,
-                    Foreground = cellColor
-                };
-                Grid.SetRow(slotCell, gridRow);
-                Grid.SetColumn(slotCell, col + 1);
-                _spellSlotTableGrid.Children.Add(slotCell);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Updates the known spells list panel below the spell slot table.
-    /// Shows spell names grouped by class and level for ALL caster classes.
-    /// Highlights overlapping spells that appear in multiple classes.
-    /// </summary>
-    private void UpdateKnownSpellsList()
-    {
-        if (_knownSpellsListPanel == null || _knownSpellsListBorder == null) return;
-
-        _knownSpellsListPanel.Children.Clear();
-
-        if (_currentCreature == null || _displayService == null)
-        {
-            _knownSpellsListBorder.IsVisible = false;
-            return;
-        }
-
-        // Get theme-aware font sizes
-        var normalFontSize = this.FindResource("FontSizeNormal") as double? ?? 14;
-        var smallFontSize = this.FindResource("FontSizeSmall") as double? ?? 12;
-        var xsmallFontSize = this.FindResource("FontSizeXSmall") as double? ?? 10;
-
-        // First pass: collect all known spells and count occurrences across classes
-        var spellOccurrences = new Dictionary<int, List<string>>(); // spellId -> list of class names
-        var casterClasses = new List<(int classIndex, string className, CreatureClass classEntry)>();
-
-        for (int i = 0; i < _currentCreature.ClassList.Count; i++)
-        {
-            var classEntry = _currentCreature.ClassList[i];
-            var className = _displayService.GetClassName(classEntry.Class) ?? $"Class {classEntry.Class}";
-
-            bool hasSpells = classEntry.KnownSpells.Any(list => list.Count > 0);
-            if (!hasSpells) continue;
-
-            casterClasses.Add((i, className, classEntry));
-
-            // Track spell occurrences
-            for (int level = 0; level <= 9; level++)
-            {
-                foreach (var spell in classEntry.KnownSpells[level])
-                {
-                    if (!spellOccurrences.ContainsKey(spell.Spell))
-                        spellOccurrences[spell.Spell] = new List<string>();
-                    spellOccurrences[spell.Spell].Add(className);
-                }
-            }
-        }
-
-        if (casterClasses.Count == 0)
-        {
-            _knownSpellsListBorder.IsVisible = false;
-            return;
-        }
-
-        bool hasAnySpells = false;
-
-        // Build spell list grouped by class, then by level
-        foreach (var (classIndex, className, classEntry) in casterClasses)
-        {
-            bool classHasSpells = false;
-
-            // Check if this class has any spells
-            for (int level = 0; level <= 9; level++)
-            {
-                if (classEntry.KnownSpells[level].Count > 0)
-                {
-                    classHasSpells = true;
-                    break;
-                }
-            }
-
-            if (!classHasSpells) continue;
-
-            hasAnySpells = true;
-            var isSelectedClass = classIndex == _selectedClassIndex;
-
-            // Class header
-            var classHeader = new TextBlock
-            {
-                Text = className,
-                FontWeight = Avalonia.Media.FontWeight.Bold,
-                FontSize = smallFontSize,
-                Foreground = isSelectedClass
-                    ? GetInfoBrush()
-                    : this.FindResource("SystemControlForegroundBaseHighBrush") as IBrush ?? GetDisabledBrush(),
-                Margin = new Avalonia.Thickness(0, hasAnySpells && _knownSpellsListPanel.Children.Count > 0 ? 12 : 0, 0, 6)
-            };
-            _knownSpellsListPanel.Children.Add(classHeader);
-
-            // Spells grouped by level
-            for (int level = 0; level <= 9; level++)
-            {
-                var knownAtLevel = classEntry.KnownSpells[level];
-                if (knownAtLevel.Count == 0) continue;
-
-                // Level header
-                var levelHeader = new TextBlock
-                {
-                    Text = level == 0 ? "  Cantrips" : $"  Level {level}",
-                    FontWeight = Avalonia.Media.FontWeight.SemiBold,
-                    FontSize = xsmallFontSize,
-                    Foreground = GetDisabledBrush(),
-                    Margin = new Avalonia.Thickness(0, 4, 0, 2)
-                };
-                _knownSpellsListPanel.Children.Add(levelHeader);
-
-                // Spell names
-                foreach (var spell in knownAtLevel)
-                {
-                    var spellName = _displayService.GetSpellName(spell.Spell);
-                    var spellInfo = _displayService.GetSpellInfo(spell.Spell);
-                    int classSpellLevel = spellInfo?.GetLevelForClass(classEntry.Class) ?? -1;
-
-                    // Check if spell appears in multiple classes (overlap)
-                    bool isOverlap = spellOccurrences.TryGetValue(spell.Spell, out var occurrences) && occurrences.Count > 1;
-
-                    // Show spell name with indicators
-                    var displayName = spellName;
-                    IBrush foreground;
-
-                    if (classSpellLevel < 0)
-                    {
-                        // Not a standard class spell (e.g., feat-based ability)
-                        displayName = $"{spellName} *";
-                        foreground = GetDisabledBrush();
-                    }
-                    else if (isOverlap)
-                    {
-                        // Spell appears in multiple classes - highlight in gold
-                        displayName = $"{spellName} ⬥";
-                        foreground = GetSelectionBrush();
-                    }
-                    else
-                    {
-                        foreground = this.FindResource("SystemControlForegroundBaseHighBrush") as IBrush ?? GetDisabledBrush();
-                    }
-
-                    var spellLabel = new TextBlock
-                    {
-                        Text = $"    {displayName}",
-                        FontSize = smallFontSize,
-                        Foreground = foreground,
-                        Margin = new Avalonia.Thickness(0, 1, 0, 1)
-                    };
-                    _knownSpellsListPanel.Children.Add(spellLabel);
-                }
-            }
-        }
-
-        _knownSpellsListBorder.IsVisible = hasAnySpells;
-    }
-
-    private void UpdateSpellSlotSummary()
-    {
-        if (_spellSlotSummaryText == null) return;
-
-        if (_currentCreature == null || _displayService == null ||
-            _selectedClassIndex >= _currentCreature.ClassList.Count)
-        {
-            _spellSlotSummaryText.Text = "";
-            return;
-        }
-
-        var classEntry = _currentCreature.ClassList[_selectedClassIndex];
-        var slots = _displayService.GetSpellSlots(classEntry.Class, classEntry.ClassLevel);
-
-        if (slots == null)
-        {
-            _spellSlotSummaryText.Text = "No spell slots for this class";
-            return;
-        }
-
-        var summaryParts = new List<string>();
-
-        for (int level = 0; level <= 9; level++)
-        {
-            int totalSlots = slots[level];
-            if (totalSlots <= 0) continue;
-
-            // Count known spells at this level
-            var knownAtLevel = classEntry.KnownSpells[level];
-            int usedSlots = knownAtLevel.Count;
-
-            // Get spell names for this level
-            var spellNames = new List<string>();
-            foreach (var spell in knownAtLevel)
-            {
-                var name = _displayService.GetSpellName(spell.Spell);
-                if (!string.IsNullOrEmpty(name))
-                    spellNames.Add(name);
-            }
-
-            // Format: "Lvl 3: 2/3 (Fireball, Haste)" or "Lvl 3: 0/3"
-            var levelSummary = $"Lvl {level}: {usedSlots}/{totalSlots}";
-            if (spellNames.Count > 0)
-            {
-                // Truncate long lists
-                var displayNames = spellNames.Count > 3
-                    ? string.Join(", ", spellNames.Take(3)) + $" +{spellNames.Count - 3}"
-                    : string.Join(", ", spellNames);
-                levelSummary += $" ({displayNames})";
-            }
-
-            summaryParts.Add(levelSummary);
-        }
-
-        if (summaryParts.Count == 0)
-        {
-            _spellSlotSummaryText.Text = "No spell slots available at current level";
-        }
-        else
-        {
-            _spellSlotSummaryText.Text = string.Join(" | ", summaryParts);
-        }
     }
 
     public void ClearPanel()
